+++ packages/admin-color-picker/src/core/HexInput.tsx
@@ -1,8 +1,7 @@
 import * as React from "react";
 import { EditableInput } from "react-color/lib/components/common";
-import * as tinycolor from "tinycolor2";
 
-import { IVPAdminColorPickerProps } from "./ColorPicker";
+import { ColorPickerProps } from "./ColorPicker";
 
 interface IComponentProps {
     value: string;
@@ -11,11 +10,7 @@ interface IComponentProps {
     palette: boolean;
 }
 
-interface IPickedColorProps {
-    value: string;
-}
-
-const resetedInputStyles = {
+const resetInputStyles = {
     input: {
         border: "inherit",
         outline: "inherit",
@@ -26,31 +21,24 @@ const resetedInputStyles = {
         margin: "inherit",
         cursor: "inherit",
         width: "100%",
+        "&::-ms-clear": {
+            display: "none",
+        },
     },
 };
 
-const PickedColor: React.FC<IPickedColorProps & IVPAdminColorPickerProps> = ({ value, classes }) => (
-    <div className={classes.pickedColorWrapper}>
-        {!value && <div className={classes.noColorStroke} />}
-        <div className={classes.pickedColorIndicator} style={{ background: value ? tinycolor(value).toHexString() : undefined }} />
-    </div>
-);
-
-const HexInput: React.FC<IComponentProps & IVPAdminColorPickerProps> = ({ value, classes, onChange, picker, palette }) => (
-    <>
-        <div className={classes.inputInner}>
-            <div className={classes.inputInnerLeftContent}>
-                <PickedColor value={value} classes={classes} />
-                {!palette || (palette && picker) ? (
-                    <EditableInput style={resetedInputStyles} value={value} onChange={(colorState)=>{
-                        onChange(colorState.hex)
-                    }} />
-                ) : (
-                    <div className={classes.readOnlyInput}>{value.toUpperCase()}</div>
-                )}
-            </div>
+const HexInput: React.FC<IComponentProps & ColorPickerProps> = ({ value, classes, onChange, picker, palette }) => (
+    <div className={classes.inputInner}>
+        <div className={classes.inputInnerLeftContent}>
+            {!palette || (palette && picker) ? (
+                <EditableInput style={resetInputStyles} value={value} onChange={(colorState)=>{
+                    onChange(colorState.hex)
+                }}  />
+            ) : (
+                <div className={classes.readOnlyInput}>{value.toUpperCase()}</div>
+            )}
         </div>
-    </>
+    </div>
 );
 
 export default HexInput;