+++ packages/admin-color-picker/CHANGELOG.md
@@ -4,8 +4,17 @@ All notable changes to this project will be documented in this file. This projec
 
 ## [next]
 
+### Highlights
+
+-   Allow custom icons/adornment for color-input
+-   The clear-button is now optional (using the `showClearButton` prop)
+
 ### Incompatible Changes
 
+-   Renamed `VPAdminColorPicker` to `CometAdminColorPicker`
+-   Removed `clearButton` and `clearIcon` classes from color-picker
+    -   Using theme-augmentation the new common clear-button can now be styled with `CometAdminClearInputButton` instead of `VPAdminColorPicker`
+-   The clear-button is no longer shown by default
 -   Removed `clearButton` and `clearIcon` classes from color-picker
     -   Using theme-augmentation the new common clear-button can now be styled with `CometAdminClearInputButton` instead of `VPAdminColorPicker`